app/requirements.txt
bcrypt
fastapi
sqlalchemy
uvicorn
PyJWT
pydantic[email]